Automata v.1.3.0
It has a precious implementation of Pushdown Automatas (PDA), deterministic and indeterministic. It also have an update of the rest of the classes, with a new method called [show], that prints on console the complete automata information. It also has a slight improvement on the examples printing, to be more fancy :3